Integer to Long

+++ lib/src/main/java/org/eclipse/tsp/java/client/api/annotation/dto/GetAnnotationsRequestDto.java
@@ -23,7 +23,7 @@ public class GetAnnotationsRequestDto {
 	private QueryInterval requestedTimerange;
 
 	@JsonProperty("requested_items")
-	private List<Integer> requestedItems;
+	private List<Long> requestedItems;
 
 	@JsonProperty("requested_marker_set")
 	private String requestedMarkerSet;
