+++ lib/src/main/java/org/eclipse/tsp/java/client/api/timegraph/dto/GetTimeGraphStatesRequestDto.java
@@ -22,5 +22,5 @@ public class GetTimeGraphStatesRequestDto {
 
 	@JsonProperty("requested_items")
 	@NonNull
-	private List<Integer> requestedItems;
+	private List<Long> requestedItems;
 }
